♻️ refactor: esc 동작 커스텀 훅으로 분리
