+++ components/src/lib/useEscapeKeyClose.ts
@@ -1,0 +1,21 @@
+import { useEffect } from "react";
+
+interface useEscapeKeyCloseProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+function useEscapeKeyClose({ isOpen, onClose }: useEscapeKeyCloseProps) {
+  useEffect(() => {
+    function handleKeyDown(e: KeyboardEvent) {
+      if (e.key === "Escape") onClose();
+    }
+
+    addEventListener("keydown", handleKeyDown);
+
+    return () => {
+      removeEventListener("keydown", handleKeyDown);
+    };
+  }, [isOpen]);
+}
+export default useEscapeKeyClose;
